tools/theme-generator/requirements.txt
google-genai>=1.0
Pillow>=10.0
numpy>=1.24
pytest>=7.0
jinja2>=3.1
